Rename `Wallet` to `WalletButton` plus move `WalletModal` outside of `shared` folder
